+++ src/sdk/python/rtdip_sdk/data_models/transformers/london_smart_meter_transformer_2_usage.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ..utils import transformers
+from ..meters.utils import transformers
 
 import datetime
 import hashlib
