--- src/sdk/python/rtdip_sdk/data_models/usage/ModelType.py
@@ -1,5 +1,0 @@
-from enum import IntFlag, auto
-
-
-class ModelType(IntFlag):
-    default = auto()
