--- src/sdk/python/rtdip_sdk/data_models/usage/UomUsage.py
@@ -1,11 +1,0 @@
-from enum import Enum
-
-
-class UomUsage(Enum):
-    w = 0
-    wh = 1
-    kw = 2
-    kwh = 3
-    mw = 4
-    mwh = 5
-    # Other
