--- src/sdk/python/rtdip_sdk/data_models/usage/pjm_voltage_lvl.py
@@ -1,25 +1,0 @@
-# Copyright 2022 RTDIP
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-from enum import Enum
-
-
-class PjmVoltagelvl(Enum):
-    transmission = 0
-    sub_transmission = 1
-    primary_substation = 2
-    primary = 3
-    secondary = 4
-    unmetered = 5
