+++ src/sdk/python/rtdip_sdk/data_models/weather/constants.py
