+++ tests/sdk/python/rtdip_sdk/data_models/utils/test_transformers.py
@@ -17,8 +17,8 @@ import sys
 sys.path.insert(0, '.')
 
 
-from src.sdk.python.rtdip_sdk.data_models.utils import transformers
-from src.sdk.python.rtdip_sdk.data_models.utils import utils
+from src.sdk.python.rtdip_sdk.data_models.meters.utils import transformers
+from src.sdk.python.rtdip_sdk.data_models.meters.utils import utils
 from datetime import date
 import datetime
